backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.13.4",
    "gspread>=6.2.1",
    "resend>=2.26.0",
    "google-auth>=2.49.1",
    "httpx>=0.27.0",
    "python-dotenv>=1.2.2",
    "fastapi>=0.135.3",
    "uvicorn>=0.43.0",
    "langchain==0.3.0",
    "langchain-core==0.3.0",
    "langchain-groq==0.2.0",
]
